Add clear examples of  jsonquery to README.md
Currently the readme shows a simple example of iterating over a json structure. But jsonquery really shines by allowing you to use an xpath like syntax to access specific keys of a map.

I've added Go playground links to each example to allow others to play around with jsonquery and try it themselves.